change license to LGPL
Due to the bzip3 third party library is
licensed under LGPL, and this crate provides
a "bundled" feature, which compiles
and statically links
the third party library.

I know for Rust
libraries, LGPL is a problem (because of
the way Cargo works). But so far
this crate is just for my own use,
and there
hasn't been so many non-LGPL-license
needs, I just don't care and let
this crate under LGPL.

version bumped